refactor: created own file for each constant
